chore: upstream
tauri: c94e132
wry: f76568a
tao: 6a03847

+++ src/millennium-core/src/platform_impl/windows/menu.rs
@@ -426,7 +426,7 @@ impl Accelerator {
 		};
 
 		Some(ACCEL {
-			fVirt: virt_key as u8,
+			fVirt: virt_key,
 			key: raw_key,
 			cmd: menu_id
 		})
